+++ examples/devenv-module/devenv.nix
@@ -29,7 +29,7 @@
 
         # Corresponding options:
         # {
-          buildType = "Debug";
+          profiles.settings.build_type = "Debug";
           compilerCppStd = "14";
 
           profiles.platformToolRequires = {