+++ examples/llvm-flake-parts/flake.nix
@@ -28,7 +28,7 @@
         in
         {
           conan = {
-            buildType = "Release";
+            profiles.settings.build_type = "Release";
             compilerCppStd = "23";
 
             profiles = {