+++ examples/standalone/flake.nix
@@ -19,7 +19,7 @@
 
             modules = [
               ({ pkgs, config, ... }: {
-                buildType = "Release";
+                profiles.settings.build_type = "Release";
                 compilerCppStd = "17";
 
                 # This should be set whenever CMakeToolchain is being used and