+++ test/flake-parts-override-default-cmd/flake.nix
@@ -25,7 +25,12 @@
           inherit (pkgs.lib) escapeShellArg;
           configLocal = "CONFIGLOCAL";
           conanHome = "./CONANHOME";
-          buildType = "Release";
+          profiles = {
+            settings = {
+              build_type = "Release";
+              os = null;
+            };
+          };
           compilerCppStd = "14";
           compilerLibCxx = "libstdc++11";
           overridingCmd = pkgs.writeShellApplication {
@@ -37,12 +42,13 @@
               echo overridingCmd "$@"
             '';
           };
+          cfg = config.conan;
         in
         {
           conan = {
             defaults.enable = true;
 
-            inherit configLocal conanHome buildType compilerCppStd compilerLibCxx;
+            inherit configLocal conanHome profiles compilerCppStd compilerLibCxx;
 
             package = overridingCmd;
 
@@ -58,7 +64,7 @@
           checks = {
             testConfigurationPackage =
               let
-                configuration = config.conan.outputs.packages.configuration;
+                configuration = cfg.outputs.packages.configuration;
                 stdenv = pkgs.gccStdenv;
                 backendStdenv = pkgs.cudaPackages.backendStdenv;
                 llvmPackages = pkgs.llvmPackages;
@@ -74,16 +80,27 @@
 
                   cat "${configuration}/.conanrc" | grep -F "conan_home="${escapeShellArg conanHome}
 
-                  cat "${configuration}/config/settings_user.yml" | grep -F ${escapeShellArg stdenv.cc.version}
-                  cat "${configuration}/config/settings_user.yml" | grep -F ${escapeShellArg backendStdenv.cc.version}
-                  cat "${configuration}/config/settings_user.yml" | grep -F ${escapeShellArg llvmPackages.libcxxStdenv.cc.version}
+                  cat "${configuration}/config/settings_user.yml" \
+                    | grep -F ${escapeShellArg stdenv.cc.version}
+                  cat "${configuration}/config/settings_user.yml" \
+                    | grep -F ${escapeShellArg backendStdenv.cc.version}
+                  cat "${configuration}/config/settings_user.yml" \
+                    | grep -F ${escapeShellArg llvmPackages.libcxxStdenv.cc.version}
 
-                  cat "${configuration}/config/profiles/default" | grep -F "build_type="${escapeShellArg buildType}
-                  cat "${configuration}/config/profiles/default" | grep -F "compiler.cppstd="${escapeShellArg compilerCppStd}
-                  cat "${configuration}/config/profiles/default" | grep -F "compiler.libcxx="${escapeShellArg compilerLibCxx}
+                  cat "${configuration}/config/profiles/default" \
+                    | grep -F "build_type="${escapeShellArg profiles.settings.build_type}
+                  cat "${configuration}/config/profiles/default" \
+                    | grep -F "compiler.cppstd="${escapeShellArg compilerCppStd}
+                  cat "${configuration}/config/profiles/default" \
+                    | grep -F "compiler.libcxx="${escapeShellArg compilerLibCxx}
 
-                  cat "${configuration}/config/profiles/default" | grep -F "[platform_tool_requires]"
-                  cat "${configuration}/config/profiles/default" | grep -F "cmake/"${escapeShellArg config.conan.final.profiles.platformToolRequires.cmake}
+                  ! cat "${configuration}/config/profiles/default" \
+                    | grep -F "os="
+
+                  cat "${configuration}/config/profiles/default" \
+                    | grep -F "[platform_tool_requires]"
+                  cat "${configuration}/config/profiles/default" \
+                    | grep -F "cmake/"${escapeShellArg cfg.final.profiles.platformToolRequires.cmake}
 
                   touch $out
                   )
@@ -91,7 +108,6 @@
 
             testLocalSetup =
               let
-                cfg = config.conan;
                 stdenv = pkgs.gccStdenv;
                 backendStdenv = pkgs.cudaPackages.backendStdenv;
                 llvmPackages = pkgs.llvmPackages;
@@ -100,6 +116,7 @@
                 {
                   name = "flake-parts-override-default-cmd-test-local-setup";
                   inherit (cfg) stdenv;
+                  derivationArgs = { inherit (cfg.outputs.devShell) buildInputs nativeBuildInputs; };
                 }
                 ''
                   (
@@ -108,34 +125,42 @@
 
                   echo "Checking local setup..."
 
-                  ${config.conan.outputs.devShell.shellHook}
+                  ${cfg.outputs.devShell.shellHook}
 
                   cat ".conanrc" | grep -F "conan_home="${escapeShellArg cfg.conanHome}
 
-                  cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" | grep -F ${escapeShellArg stdenv.cc.version}
-                  cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" | grep -F ${escapeShellArg backendStdenv.cc.version}
-                  cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" | grep -F ${escapeShellArg llvmPackages.libcxxStdenv.cc.version}
+                  cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
+                    | grep -F ${escapeShellArg stdenv.cc.version}
+                  cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
+                    | grep -F ${escapeShellArg backendStdenv.cc.version}
+                  cat ${escapeShellArg cfg.configLocal}"/settings_user.yml" \
+                    | grep -F ${escapeShellArg llvmPackages.libcxxStdenv.cc.version}
 
-                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" | grep -F "build_type="${escapeShellArg cfg.buildType}
-                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" | grep -F "compiler.cppstd="${escapeShellArg cfg.compilerCppStd}
-                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" | grep -F "compiler.libcxx="${escapeShellArg cfg.compilerLibCxx}
+                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
+                    | grep -F "build_type="${escapeShellArg cfg.final.profiles.settings.build_type}
+                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
+                    | grep -F "compiler.cppstd="${escapeShellArg cfg.compilerCppStd}
+                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
+                    | grep -F "compiler.libcxx="${escapeShellArg cfg.compilerLibCxx}
 
-                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" | grep -F "[platform_tool_requires]"
-                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" | grep -F "cmake/"${escapeShellArg cfg.final.profiles.platformToolRequires.cmake}
+                  ! cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
+                    | grep -F "os="
+
+                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
+                    | grep -F "[platform_tool_requires]"
+                  cat ${escapeShellArg cfg.configLocal}"/profiles/default" \
+                    | grep -F "cmake/"${escapeShellArg cfg.final.profiles.platformToolRequires.cmake}
 
                   touch $out
                   )
                 '';
 
             testConanPackage =
-              let
-                cfg = config.conan;
-              in
               pkgs.runCommandWith
                 {
                   name = "flake-parts-override-default-cmd-test-conan-profile";
                   inherit (cfg) stdenv;
-                  derivationArgs = { inherit (config.conan.outputs.devShell) buildInputs nativeBuildInputs; };
+                  derivationArgs = { inherit (cfg.outputs.devShell) buildInputs nativeBuildInputs; };
                 }
                 ''
                   (
@@ -144,7 +169,7 @@
 
                   echo "Checking Conan is not on path..."
 
-                  ${config.conan.outputs.devShell.shellHook}
+                  ${cfg.outputs.devShell.shellHook}
 
                   echo "Package: "${getCommand cfg.package}