+++ test/llvm-flake-parts/flake.nix
@@ -17,7 +17,7 @@
       ];
       perSystem = { self', pkgs, config, ... }: {
         conan = {
-          buildType = "Release";
+          profiles.settings.build_type = "Release";
           compilerCppStd = "23";
           stdenv = pkgs.overrideCC
             (